+++ packages/process/lib/Polysemy/Process/Interpreter/Process.hs
@@ -4,6 +4,7 @@
 module Polysemy.Process.Interpreter.Process where
 
 import Control.Concurrent.STM.TBMQueue (TBMQueue)
+import Data.ByteString (hGetSome, hPut)
 import qualified Polysemy.Conc as Conc
 import Polysemy.Conc.Async (withAsync_)
 import qualified Polysemy.Conc.Data.QueueResult as QueueResult
@@ -13,8 +14,11 @@ import Polysemy.Conc.Effect.Race (Race)
 import Polysemy.Conc.Effect.Scoped (Scoped)
 import Polysemy.Conc.Interpreter.Queue.TBM (interpretQueueTBMWith, withTBMQueue)
 import Polysemy.Conc.Interpreter.Scoped (interpretScopedResumableWith_)
-import Polysemy.Resume (Stop, resumeOr, resume_, stop, type (!!))
+import Polysemy.Input (Input (Input))
+import Polysemy.Output (Output (Output))
+import Polysemy.Resume (Stop, interpretResumable, resumeOr, resume_, stop, stopNote, type (!!))
 import Prelude hiding (fromException)
+import System.IO (BufferMode (NoBuffering), Handle, hSetBuffering, stdin, stdout)
 
 import Polysemy.Process.Data.ProcessError (ProcessError (Terminated))
 import Polysemy.Process.Data.ProcessKill (ProcessKill (KillAfter, KillImmediately, KillNever))
@@ -92,6 +96,7 @@ withSTMResources qSize action = do
       action (ProcessQueues inQ outQ)
 
 withQueues ::
+  ∀ i o r .
   Members [Race, Resource, Embed IO] r =>
   Int ->
   InterpretersFor [Queue (In i), Queue (Out o)] r
@@ -99,10 +104,10 @@ withQueues qSize action =
   withSTMResources qSize \ qs -> interpretQueues qs action
 
 outputQueue ::
-  ∀ p chunk err r .
-  Members [SystemProcess !! err, ProcessOutput p chunk, Queue (Out chunk), Embed IO] r =>
+  ∀ p chunk err eff r .
+  Members [eff !! err, ProcessOutput p chunk, Queue (Out chunk), Embed IO] r =>
   Bool ->
-  Sem (SystemProcess : r) ByteString ->
+  Sem (eff : r) ByteString ->
   Sem r ()
 outputQueue discardWhenFull readChunk = do
   spin ""
@@ -116,9 +121,9 @@ outputQueue discardWhenFull readChunk = do
       spin newBuffer
 
 inputQueue ::
-  ∀ i err r .
-  Members [SystemProcess !! err, ProcessInput i, Queue (In i), Embed IO] r =>
-  (ByteString -> Sem (SystemProcess : r) ()) ->
+  ∀ i err eff r .
+  Members [eff !! err, ProcessInput i, Queue (In i), Embed IO] r =>
+  (ByteString -> Sem (eff : r) ()) ->
   Sem r ()
 inputQueue writeChunk =
   spin
@@ -166,9 +171,9 @@ scope ::
 scope (ProcessOptions discard qSize kill) =
   withSystemProcess @resource .
   withQueues qSize .
-  withAsync_ (outputQueue @'Stderr @o @err discard SystemProcess.readStderr) .
-  withAsync_ (outputQueue @'Stdout @o @err discard SystemProcess.readStdout) .
-  withAsync_ (inputQueue @i @err SystemProcess.writeStdin) .
+  withAsync_ (outputQueue @'Stderr @o @err @SystemProcess discard SystemProcess.readStderr) .
+  withAsync_ (outputQueue @'Stdout @o @err @SystemProcess discard SystemProcess.readStdout) .
+  withAsync_ (inputQueue @i @err @SystemProcess SystemProcess.writeStdin) .
   withKill @err kill
 
 -- |Interpret 'Process' with a system process resource whose file descriptors are connected to three 'TBMQueue's,
@@ -250,3 +255,89 @@ interpretInputOutputProcess ::
 interpretInputOutputProcess =
   runOutputSem (Process.send @i @o) .
   runInputSem (Process.recv @i @o)
+
+-- |Interpret 'Input ByteString' by polling a 'Handle' and stopping with 'ProcessError' when it fails.
+interpretInputHandleBuffered ::
+  Member (Embed IO) r =>
+  Handle ->
+  InterpreterFor (Input ByteString !! ProcessError) r
+interpretInputHandleBuffered handle =
+  interpretResumable \case
+    Input ->
+      stopNote (Terminated "handle closed") =<< tryMaybe (hGetSome handle 4096)
+
+-- |Interpret 'Input ByteString' by polling a 'Handle' and stopping with 'ProcessError' when it fails.
+-- This variant deactivates buffering for the 'Handle'.
+interpretInputHandle ::
+  Member (Embed IO) r =>
+  Handle ->
+  InterpreterFor (Input ByteString !! ProcessError) r
+interpretInputHandle handle sem = do
+  void $ tryMaybe (hSetBuffering handle NoBuffering)
+  interpretInputHandleBuffered handle sem
+
+-- |Interpret 'Output ByteString' by writing to a 'Handle' and stopping with 'ProcessError' when it fails.
+interpretOutputHandleBuffered ::
+  Member (Embed IO) r =>
+  Handle ->
+  InterpreterFor (Output ByteString !! ProcessError) r
+interpretOutputHandleBuffered handle =
+  interpretResumable \case
+    Output o ->
+      stopNote (Terminated "handle closed") =<< tryMaybe (hPut handle o)
+
+-- |Interpret 'Output ByteString' by writing to a 'Handle' and stopping with 'ProcessError' when it fails.
+-- This variant deactivates buffering for the 'Handle'.
+interpretOutputHandle ::
+  Member (Embed IO) r =>
+  Handle ->
+  InterpreterFor (Output ByteString !! ProcessError) r
+interpretOutputHandle handle sem = do
+  void $ tryMaybe (hSetBuffering handle NoBuffering)
+  interpretOutputHandleBuffered handle sem
+
+-- |Interpret 'Process' in terms of 'Input' and 'Output'.
+-- Since the @i@ and @o@ parameters correspond to the abstraction of stdio fds of an external system process, @i@ is
+-- written by 'Output' and @o@ is read from 'Input'.
+-- This is useful to abstract the current process's stdio as an external process, with input and output swapped.
+interpretProcessIO ::
+  ∀ i o ie oe r .
+  Members [Input ByteString !! ie, Output ByteString !! oe] r =>
+  Members [ProcessInput i, ProcessOutput 'Stdout o, Resource, Race, Async, Embed IO] r =>
+  ProcessOptions ->
+  InterpreterFor (Process i o !! ProcessError) r
+interpretProcessIO (ProcessOptions discard qSize _) =
+  withQueues @i @o qSize .
+  withAsync_ (outputQueue @'Stdout @o @ie @(Input ByteString) discard input) .
+  withAsync_ (inputQueue @i @oe @(Output ByteString) output) .
+  interpretResumable handleProcessWithQueues .
+  raiseUnder2
+
+-- |Interpret 'Process' in terms of two 'Handle's.
+-- This is useful to abstract the current process's stdio as an external process, with input and output swapped.
+-- The first 'Handle' argument corresponds to the @o@ parameter, the second one to @i@, despite the first one usually
+-- being the current process's stdin.
+-- This is due to 'Process' abstracting an external process to whose stdin would be /written/, while the current one's
+-- is /read/.
+interpretProcessHandles ::
+  ∀ i o r .
+  Members [ProcessInput i, ProcessOutput 'Stdout o, Resource, Race, Async, Embed IO] r =>
+  ProcessOptions ->
+  Handle ->
+  Handle ->
+  InterpreterFor (Process i o !! ProcessError) r
+interpretProcessHandles options hIn hOut =
+  interpretOutputHandle hOut .
+  interpretInputHandle hIn .
+  interpretProcessIO @i @o @ProcessError @ProcessError options .
+  raiseUnder2
+
+-- |Interpret 'Process' using the current process's stdin and stdout.
+-- This mirrors the usual abstraction of an external process, to whose stdin would be /written/, while the current one's
+-- is /read/.
+interpretProcessCurrent ::
+  Members [ProcessInput i, ProcessOutput 'Stdout o, Resource, Race, Async, Embed IO] r =>
+  ProcessOptions ->
+  InterpreterFor (Process i o !! ProcessError) r
+interpretProcessCurrent options =
+  interpretProcessHandles options stdin stdout